Create Scan tab and adapt old Measurement tab
Create a tab for scan measurement and adapt the old measurement tab
to be a "position control and manual measurement verification" tab,
named only as "Control".

Following the logic of scanning a sequence of magnet blocks,
even single block measurements could be done in the scan tab,
leaving the "Control" tab for manual position adjustment and
verifying measurements without the need to store them.