Add option to hide error/ sample data in scan tab
Since the error range could be a lot smaller than the
sample values, this commit adds an option to hide sample
data and autoscale graph allowing error data to be better
visualized in the scan tab.

An alternative would be to add another scale to the plot.
At the moment there is no straightforward approach for
doing that with the PyQt PlotWidget, so it is left as a
possible future improvement, since the current solution
meets the application needs.