Add check for max speed and number of points per scan
Add check for max number of points at the beginning of a hall sensor
scan, so that the maximum number of readings that can be stored by
the multimeter is not exceeded.

More points can be sampled by dividing the scan in multiple sub-scans,
which, after they are stored in the database, can be displayed together
to the user on the GUI graph.