Do not allow saving new scan data that overlaps with previous data
If a new scan has data that completly or partially overlaps with
already stored data for the same scan ID (same measurement,
undulator and cassette names) then an attempt to save the new
data fails and an error message box is shown to the user.

In order to overwrite the old data, it must first be deleted
through the database tab.