Fix function to fetch readings from multimeter
The previous implementation would result in problems when
the number of readings to fetch was relatively large (roughly
greater than 500).

Reading the serial input periodically and concatenating the
input until the message was complete, instead of just waiting
a long time after the fetch command to read solved the problem.