Make single multimeter measurement more robust
Make the single Hall measurement available at the
Control tab more robust.

Sending a "device clear" (Ctrl-C) to the instrument
before taking a single measurement guarantees that the
device is not waiting for a trigger, which is a condition
that could arise in case some misconfiguration or failure
happened at some moment before.