Add skeleton files for the assembly tab
A new feature of the Delta Assembly Bench will
be a tab to instruct the user on how to assemble
the device magnet blocks, according to the information
of the specified input file.

This commit adds basic files to serve as a skeleton for
the .ui and tab actual code.