Add independent horizontal velocity for Assembly tab
Add double spin boxes for horizontal velocity and acceleration
on the Assembly tab. Since the manual motion controls available
in this tab are not necessarily related to the motion parameters
used for measurements, it is interesting that different fields are
used to define velocity and acceleration there, instead of reeading
the configurations found in the advanced options dialog (which are
used for scan and individual measurement).