Add encoder reading display to assembly tab
Implement encoder reading display in assembly tab
by sharing the readings obtained periodically by the
Control tab (MeasurementWidget) with the assembly
tab (AssemblyWidget). This is done through global
variables in the Main App (mainapp.py).

The Scan tab (ScanWidget) disables the encoder periodic
reading fetching done by the Control tab when it starts
a scan. It does so by setting the corresponding global
variable (encoder_update_enabled) in the main app to false.
After the scan is complete or canceled, the encoder update
enabled variable is restored to true.