scanwidget: Add button to save scan data to excel file
Previously, it was possible to save all the data gathered so far for
a given scan to an excel file when new data was being saved to the database,
as the user was asked whether the complete data should also be written
to excel.

There would be a problem if the user did not accept the excel file to
be written at database update, but then later wanted to create the excel
file.

This commit fix this by adding a new button to the UI which allows the
user, at any time, to save all accumulated data for the scan specified
in the name boxes (measurement, undulator and cassette) to an excel
workbook.